Add files via upload
Airbnb is a popular online platform that allows individuals to rent out their properties or spare rooms to travelers. Exploring Airbnb data in SQL can provide insights into various aspects of the platform, including pricing, location, and availability.
One of our hypothesis going into the exploration was the expectation of a variable we titled number of amenities to have a considerable effect on the price of property listings, we found that this was not the case.

In our exploration of Airbnb data in SQL, we may expect to find:

Price variability: We can analyze the pricing of different properties based on their location, type, and amenities. We may expect to find that certain areas or property types have higher average prices than others.

Booking trends: We can examine the frequency and timing of bookings to determine popular travel periods, seasonal trends, and high-demand locations. We may expect to find that certain cities or neighborhoods have higher booking rates during specific times of the year.

Property characteristics: We can investigate the characteristics of different Airbnb properties, such as their size, amenities, and ratings. We may expect to find that certain amenities, such as a pool or a private balcony, increase the likelihood of bookings.

User behavior: We can analyze user behavior on the platform, such as the number of reviews left and the frequency of bookings. We may expect to find that users who leave more reviews are more likely to be repeat customers.

Overall, exploring Airbnb data in SQL can provide valuable insights into the behavior of both hosts and guests on the platform, which can help inform pricing, marketing, and other business decisions.